refactor(chat): relocate the editing context into the chat package and drop MessageGroup's dead provider (A2)
The message-editing context lived at @renderer/context — an app-level location that
made it a cross-cutting dependency outside the chat package. Move it to
components/chat/messages/editing/ so it is chat-owned and migrates with the chat tree
rather than as a separate app dependency. Now that MessageFrame reads editing state
through the MessageList provider interface (A1), MessageGroup's nested
MessageEditingProvider is dead — remove it. ChatComposer keeps consuming the (now
chat-internal) context, which is ordinary intra-package lifted-state coordination, not
the leaf-reaches-app-context inversion that A1 removed.
